Fix amalgam average charge
An amalgam is assumed to be a mixture of elements, so it won't have an
integral value of Z.  This changes the scheme for an amalgam to expect
a floating point for Z (instead of an integer).

+++ python/gegede/schema/__init__.py
@@ -64,7 +64,7 @@ Schema = dict(
         Composition = (("symbol",str), ("isotopes", NamedTypedList(float))),
 
         # a material with no specific constituents
-        Amalgam = (("z", int),
+        Amalgam = (("z", float),
                    ("a","0.0g/mole"),
                    ("density", "0.0g/cc"), 
                    ("properties",NamedTypedList(list))),
